+++ common/src/main/java/net/mehvahdjukaar/supplementaries/mixins/AbstractSkeletonMixin.java
@@ -36,7 +36,7 @@ public abstract class AbstractSkeletonMixin extends Monster {
     public void supp$finalizeSpawn(ServerLevelAccessor level, DifficultyInstance difficulty, MobSpawnType spawnType, SpawnGroupData spawnGroupData, CallbackInfoReturnable<SpawnGroupData> cir) {
         if (this.getType() == EntityType.SKELETON || this.getType() == EntityType.STRAY && CommonConfigs.Tools.QUIVER_ENABLED.get()) {
             double chance = CommonConfigs.Tools.QUIVER_SKELETON_SPAWN.get() * (
-                    CommonConfigs.Tools.QUIVER_DEPEND_ON_GLOBAL_DIFFICULTY.get() ? difficulty.getSpecialMultiplier() :1);
+                    CommonConfigs.Tools.QUIVER_DEPEND_ON_GLOBAL_DIFFICULTY.get() ? difficulty.getSpecialMultiplier() :0.1);
             if (random.nextFloat() < chance) {
                 ((IQuiverEntity) this).supplementaries$setQuiver(
                         RandomArrowFunction.createRandomQuiver(level.getRandom(), difficulty.getSpecialMultiplier()));