+++ common/src/main/java/net/mehvahdjukaar/supplementaries/reg/LootTablesInjects.java
@@ -10,6 +10,7 @@ import net.minecraft.world.level.storage.loot.BuiltInLootTables;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
+import java.util.SimpleTimeZone;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
@@ -61,7 +62,6 @@ public class LootTablesInjects {
         PYRAMID,
         STRONGHOLD,
         TEMPLE,
-        TEMPLE_DISPENSER,
         IGLOO,
         MANSION,
         FORTRESS,
@@ -81,7 +81,6 @@ public class LootTablesInjects {
         if (isMineshaft(name)) return TableType.MINESHAFT;
         if (isDungeon(name)) return TableType.DUNGEON;
         if (isTemple(name)) return TableType.TEMPLE;
-        if (isTempleDispenser(name)) return TableType.TEMPLE_DISPENSER;
         if (isOutpost(name)) return TableType.PILLAGER;
         if (isStronghold(name)) return TableType.STRONGHOLD;
         if (isFortress(name)) return TableType.FORTRESS;
@@ -91,60 +90,68 @@ public class LootTablesInjects {
         return TableType.OTHER;
     }
 
+    private static final String FISHING_TREASURE = BuiltInLootTables.FISHING_TREASURE.location().toString();
     private static boolean isFishTreasure(String name) {
-        return name.equals(BuiltInLootTables.FISHING_TREASURE.toString());
+        return name.equals(FISHING_TREASURE);
     }
 
+    private static final String MANSION = BuiltInLootTables.WOODLAND_MANSION.location().toString();
     private static boolean isMansion(String name) {
-        return name.equals(BuiltInLootTables.WOODLAND_MANSION.toString()) || RS && name.contains("repurposed_structures:chests/mansion");
+        return name.equals(MANSION) || RS && name.contains("repurposed_structures:chests/mansion");
     }
 
     private static final Pattern RS_SHIPWRECK = Pattern.compile("repurposed_structures:chests/shipwreck/\\w*/treasure_chest");
-
+    private static final String SHIPWRECK_TREASURE = BuiltInLootTables.SHIPWRECK_TREASURE.location().toString();
     private static boolean isShipwreck(String s) {
-        return s.equals(BuiltInLootTables.SHIPWRECK_TREASURE.toString()) || RS && RS_SHIPWRECK.matcher(s).matches();
+        return s.equals(SHIPWRECK_TREASURE) || RS && RS_SHIPWRECK.matcher(s).matches();
     }
 
     private static final Pattern RS_SHIPWRECK_STORAGE = Pattern.compile("repurposed_structures:chests/shipwreck/\\w*/supply_chest");
-
+    private static final String SHIPWRECK_SUPPLY = BuiltInLootTables.SHIPWRECK_SUPPLY.location().toString();
     private static boolean isShipwreckStorage(String s) {
-        return s.equals(BuiltInLootTables.SHIPWRECK_SUPPLY.toString()) || RS && RS_SHIPWRECK_STORAGE.matcher(s).matches();
+        return s.equals(SHIPWRECK_SUPPLY) || RS && RS_SHIPWRECK_STORAGE.matcher(s).matches();
     }
 
+    private static final String ABANDONED_MINESHAFT = BuiltInLootTables.ABANDONED_MINESHAFT.location().toString();
     private static boolean isMineshaft(String s) {
-        return s.equals(BuiltInLootTables.ABANDONED_MINESHAFT.toString()) || RS && s.contains("repurposed_structures:chests/mineshaft");
+        return s.equals(ABANDONED_MINESHAFT) || RS && s.contains("repurposed_structures:chests/mineshaft");
     }
 
+    private static final String PILLAGER_OUTPOST = BuiltInLootTables.PILLAGER_OUTPOST.location().toString();
     private static boolean isOutpost(String s) {
-        return s.equals(BuiltInLootTables.PILLAGER_OUTPOST.toString()) || RS && s.contains("repurposed_structures:chests/outpost");
+        return s.equals(PILLAGER_OUTPOST) || RS && s.contains("repurposed_structures:chests/outpost");
     }
 
+    private static final String DUNGEON = BuiltInLootTables.SIMPLE_DUNGEON.location().toString();
     private static boolean isDungeon(String s) {
-        return s.equals(BuiltInLootTables.SIMPLE_DUNGEON.toString()) || RS && s.contains("repurposed_structures:chests/dungeon");
+        return s.equals(DUNGEON) || RS && s.contains("repurposed_structures:chests/dungeon");
     }
 
     private static final Pattern RS_TEMPLE = Pattern.compile("repurposed_structures:chests/temple/\\w*_chest");
-
+    private static final String JUNGLE_TEMPLE = BuiltInLootTables.JUNGLE_TEMPLE.location().toString();
     private static boolean isTemple(String s) {
-        return s.equals(BuiltInLootTables.JUNGLE_TEMPLE.toString()) || RS && RS_TEMPLE.matcher(s).matches();
+        return s.equals(JUNGLE_TEMPLE) || RS && RS_TEMPLE.matcher(s).matches();
     }
 
     private static final Pattern RS_TEMPLE_DISPENSER = Pattern.compile("repurposed_structures:chests/temple/\\w*_dispenser");
-
+    private static final String JUNGLE_TEMPLE_DISPENSER = BuiltInLootTables.JUNGLE_TEMPLE_DISPENSER.location().toString();
     private static boolean isTempleDispenser(String s) {
-        return s.equals(BuiltInLootTables.JUNGLE_TEMPLE.toString()) || RS && RS_TEMPLE_DISPENSER.matcher(s).matches();
+        return s.equals(JUNGLE_TEMPLE_DISPENSER) || RS && RS_TEMPLE_DISPENSER.matcher(s).matches();
     }
 
+    private static final String STRONGHOLD_CROSSING = BuiltInLootTables.STRONGHOLD_CROSSING.location().toString();
     private static boolean isStronghold(String s) {
-        return s.equals(BuiltInLootTables.STRONGHOLD_CROSSING.toString()) || RS && s.contains("repurposed_structures:chests/stronghold/nether_storage_room");
+        return s.equals(STRONGHOLD_CROSSING) || RS && s.contains("repurposed_structures:chests/stronghold/nether_storage_room");
     }
 
+    private static final String NETHER_BRIDGE = BuiltInLootTables.NETHER_BRIDGE.location().toString();
     private static boolean isFortress(String s) {
-        return s.equals(BuiltInLootTables.NETHER_BRIDGE.toString()) || RS && s.contains("repurposed_structures:chests/fortress");
+        return s.equals(NETHER_BRIDGE) || RS && s.contains("repurposed_structures:chests/fortress");
     }
 
+    private static final String END_CITY_TREASURE = BuiltInLootTables.END_CITY_TREASURE.location().toString();
     private static boolean isEndCity(String s) {
-        return s.equals(BuiltInLootTables.END_CITY_TREASURE.toString());
+        return s.equals(END_CITY_TREASURE);
     }
 
     private static void injectPool(Consumer<ResourceLocation> consumer, TableType type, String name) {
@@ -160,7 +167,7 @@ public class LootTablesInjects {
     }
 
     private static void tryInjectQuiver(Consumer<ResourceLocation> e, TableType type) {
-        if (type == TableType.DUNGEON || type == TableType.MANSION) {
+        if (type == TableType.DUNGEON || type == TableType.MANSION || type == TableType.PILLAGER) {
             injectPool(e, type, "quiver");
         }
     }